if read function fails

+++ includes/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: shunwata <shunwata@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/29 20:46:07 by shunwata          #+#    #+#             */
-/*   Updated: 2025/06/08 21:55:21 by shunwata         ###   ########.fr       */
+/*   Updated: 2025/06/12 18:43:17 by shunwata         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@
 
 int	ft_putchar(char c);
 int	ft_putstr(char *str);
-int	ft_putnbr_base(unsigned long long n, int base, int uppercase);
+int	ft_putnbr_base(unsigned long n, int base, int uppercase);
 int	ft_putnbr(int n);
 int	put_pointer(void *ptr);
